i2006-08-14  Mark Wielaard  <mark@klomp.org>
    * Makefile.am (ftk.jar): Add all classes to jar.
    * src/java/com/redhat/ftk/EventViewer.java (getSelectedTraces):
    Returns long[].
    (ftk_eventviewer_get_selected_traces): Likewise.
    * src/jni/com_redhat_ftk_EventViewer.c
(Java_com_redhat_ftk_EventViewer_ftk_1eventviewer_1get_1selected_1traces):
    Returns jlongArray.
    * src/widgets/ftkeventviewer.c
    (ftk_eventviewer_get_selected_traces): Uses longs.

2006-08-14  Mark Wielaard  <mark@klomp.org>

    * StatusWidget.java (StatusWidget): MouseListener.mouseEvent()
    traces[] consists of longs.

+++ frysk-gtk/tlwidgets/src/widgets/ftkeventviewer.c
@@ -3021,7 +3021,7 @@ ftk_eventviewer_set_trace_linestyle (FtkEventViewer * eventviewer,
  ftk_eventviewer_get_selected_traces (FtkEventViewer * eventviewer)
  {
  	GList * traces = NULL;
- 	for (gint traceID = 0; traceID < ftk_ev_trace_next(eventviewer); 
+ 	for (long traceID = 0; traceID < ftk_ev_trace_next(eventviewer); 
  	traceID++) 
  	{
  		FtkTrace *trace = ftk_ev_trace(eventviewer, traceID);
